+++ Code/ConfigOptions.cs
@@ -3,39 +3,43 @@ using System.Collections.Generic;
 using System.Text;
 using BepInEx.Configuration;
 using RoR2;
+using MiscFixes.Modules;
 
 namespace ShardTierClassicItems
 {
     public static class ConfigOptions
     {
         internal static ConfigEntry<bool> LogExtraInfo;
+
         internal static List<ConfigEntry<string>> TradesConfigEntriesList = [];
         internal static List<ConfigEntry<bool>> TierChangeConfigEntriesList = [];
 
         internal static void BindConfigOptions(ConfigFile config)
         {
-            LogExtraInfo = config.Bind<bool>(
+            LogExtraInfo = config.BindOption(
                 "Other",
                 "Enable logging extra info",
-                false,
-                "If you wanna see stuff like all valid trades, the mod changing item tiers and adding items to trades, enable this."
+                "If you wanna see stuff like all valid trades, the mod changing item tiers and adding items to trades, enable this.",
+                false
             );
 
-            string configOptionDescription = "Set what shard trades this item should be added to.\nValid options:\n";
+
+
+            string tradeConfigOptionDescription = "Set what shard trades this item should be added to.\nValid options:\n";
             foreach (var tradeName in Main.ValidTradeNames)
             {
                 // i could do a smart check for it if's the last value in the list but it's not needed here 
                 if (tradeName == "None")
                 {
-                    configOptionDescription += tradeName;
+                    tradeConfigOptionDescription += tradeName;
                 }
                 else
                 {
-                    configOptionDescription += $"{tradeName}, ";
+                    tradeConfigOptionDescription += $"{tradeName}, ";
                 }
             }
-            configOptionDescription += "\nSeparate each one with a COMMA ( , ) and NO SPACES, except for `All` or `None` which must be alone.";
-            configOptionDescription += "\n\nNOTE: Setting to `All` means it may add itself to other modded trades!";
+            tradeConfigOptionDescription += "\nSeparate each one with a COMMA ( , ) and NO SPACES, except for `All` or `None` which must be alone.";
+            tradeConfigOptionDescription += "\n\nNOTE: Setting to `All` means it may add itself to other modded trades!";
 
             foreach (var itemDef in RoR2.ContentManagement.ContentManager.itemDefs)
             {
@@ -47,18 +51,18 @@ namespace ShardTierClassicItems
 
                 // changing the language would cause the config to use a differently named set of options
                 // so let's stick to english item names to prevent that
-                TradesConfigEntriesList.Add(config.Bind<string>(
+                TradesConfigEntriesList.Add(config.BindOption(
                     "Add to shard trades",
                     englishItemName,
-                    "All",
-                    configOptionDescription
+                    tradeConfigOptionDescription,
+                    "All"
                 ));
 
-                TierChangeConfigEntriesList.Add(config.Bind<bool>(
+                TierChangeConfigEntriesList.Add(config.BindOption(
                     "Change to shard tier",
                     englishItemName,
-                    true,
-                    "Should this item be changed to shard tier? The item icon's outline does not yet change color to match this."
+                    "Should this item be changed to shard tier?",
+                    true
                 ));
             }
         }